Fixes Kelvin Degree as per:
Kelvin is an absolute temperature scale with a defined endpoint, and therefore doesn't use degrees1. The "degree Kelvin" (°K) is a former name and symbol for the SI unit of temperature on the thermodynamic (absolute) temperature scale. Since 1967, it has been known simply as the kelvin, with symbol K (without a degree symbol)